+++ src/SolidSphere.cpp
@@ -8,8 +8,8 @@
 SolidSphere::SolidSphere(float radius, unsigned int rings, unsigned int sectors)
 : Primitive(Primitive::Types::SOLID_SPHERE)
 {
-	float const R = 1. / (float)(rings - 1);
-	float const S = 1. / (float)(sectors - 1);
+	float const R = 1.0 / (float)(rings - 1);
+	float const S = 1.0 / (float)(sectors - 1);
 	unsigned int r, s;
 
 	vertices.resize(rings * sectors);
@@ -21,9 +21,9 @@ SolidSphere::SolidSphere(float radius, unsigned int rings, unsigned int sectors)
 	{
 		for (s = 0; s < sectors; s++)
 		{
-			float const y = sin(-pi_2 + M_PI * r * R);
-			float const x = cos(2 * M_PI * s * S) * sin(M_PI * r * R);
-			float const z = sin(2 * M_PI * s * S) * sin(M_PI * r * R);
+			float const y = (float)sin(-pi_2 + M_PI * r * R);
+			float const x = (float)cos(2 * M_PI * s * S) * sin(M_PI * r * R);
+			float const z = (float)sin(2 * M_PI * s * S) * sin(M_PI * r * R);
 
 			v->x = x * radius;
 			v->y = y * radius;
@@ -36,10 +36,10 @@ SolidSphere::SolidSphere(float radius, unsigned int rings, unsigned int sectors)
 	indices.resize(rings * sectors * 4);
 	std::vector<short>::iterator i = indices.begin();
 	for (r = 0; r < rings - 1; r++) for (s = 0; s < sectors - 1; s++) {
-		*i++ = r * sectors + s;
-		*i++ = r * sectors + (s + 1);
-		*i++ = (r + 1) * sectors + (s + 1);
-		*i++ = (r + 1) * sectors + s;
+		*i++ = (short) r * sectors + s;
+		*i++ = (short) r * sectors + (s + 1);
+		*i++ = (short) (r + 1) * sectors + (s + 1);
+		*i++ = (short) (r + 1) * sectors + s;
 	}
 
 	// Create buffer and load vertex data
